+++ fastkml/__init__.py
@@ -31,6 +31,7 @@ from fastkml.atom import Contributor as AtomContributor
 from fastkml.atom import Link as AtomLink
 from fastkml.containers import Document
 from fastkml.containers import Folder
+from fastkml.network_link_control import NetworkLinkControl
 from fastkml.data import Data
 from fastkml.data import ExtendedData
 from fastkml.data import Schema
@@ -84,7 +85,7 @@ from fastkml.views import LookAt
 
 __all__ = [
     "KML",
-    "NetworkLinkControl"
+    "NetworkLinkControl",
     "AtomAuthor",
     "AtomContributor",
     "AtomLink",
